DAO commit

+++ Trabalho POO/Models/Desconto.cs
@@ -5,7 +5,7 @@ namespace Mercado.Models
     internal class Desconto
     {
         public int Id_desconto { get; set; }
-        public decimal Porcentagem { get; set; }
+        public double Porcentagem { get; set; }
         public DateTime DataInicio { get; set; }
         public DateTime DataFim { get; set; }
     }
